+++ wsgo-lib/src/main/java/com/gnepux/wsgo/retry/RetryStrategy.java
@@ -1,7 +1,20 @@
 package com.gnepux.wsgo.retry;
 
+/**
+ * Retry strategy when reconnect.
+ *
+ * @author gnepux
+ * @see DefaultRetryStrategy
+ */
 public interface RetryStrategy {
 
-    long onRetry(int retryCount);
+    /**
+     * The relationship of retry count and delay time,
+     * WsGo will call the method for every reconnect.
+     *
+     * @param retryCount The retry time that WsGo has retied.
+     * @return The delay time in milliseconds.
+     */
+    long onRetry(long retryCount);
 
 }
